chore: rename some commands to group them a bit better
