--- aimbat/lib/plot/__init__.py
@@ -1,4 +1,0 @@
-from .seismograms import plot_seismograms
-from .stack import plot_stack
-
-__all__ = ["plot_seismograms", "plot_stack"]
