+++ tests/test_seismogram.py
@@ -60,6 +60,19 @@ class TestLibSeismogram:
                 parameter_value=False,
             )
 
+    def test_plotseis(self, test_data, db_session, mock_show) -> None:  # type: ignore
+        from aimbat.lib import data, event
+        from aimbat.lib.models import AimbatEvent
+
+        data.add_files_to_project(
+            db_session, test_data, filetype=SeismogramFileType.SAC
+        )
+
+        aimbat_event = db_session.get(AimbatEvent, 1)
+        assert aimbat_event.active is None
+        event.set_active_event(db_session, aimbat_event)
+        seismogram.plot_seismograms(db_session)
+
 
 class TestCliSeismogram:
     def test_sac_data(self, db_url, test_data_string, monkeypatch, capsys) -> None:  # type: ignore
@@ -147,3 +160,23 @@ class TestCliSeismogram:
 
         app(["event", "activate", "1", "--db-url", db_url])
         app(["seismogram", "list", "--db-url", db_url])
+
+    def test_plotseis(self, test_data_string, db_url, mock_show, capsys) -> None:  # type: ignore
+        """Test AIMBAT cli with utils subcommand."""
+
+        from aimbat.app import app
+
+        app(["utils"])
+        assert "Usage" in capsys.readouterr().out
+
+        app(["project", "create", "--db-url", db_url])
+
+        args = ["data", "add", "--db-url", db_url]
+        args.extend(test_data_string)
+        app(args)
+
+        app(["event", "activate", "1", "--db-url", db_url])
+
+        app(["seismogram", "plot", "--db-url", db_url])
+
+        # app(["seismogram", "plot", "--db-url", db_url, "--use-qt"])
